win/ntfs: remove unused variable and have ntfssect use char API calls
The variable 'ok' is never used and generates a warning. Remove it.
Also ntfssect.c is designed to be compiled in non Unicode mode when
using MSVC compilers, so remove all ambiguity about it
(LPCTSTR -> LPCSTR, use of 'A' API calls) so that it doesn't break when
compiled in Unicode mode, which is what Rufus uses with MSVC.

Signed-off-by: Pete Batard <pete@akeo.ie>